Commit before heavy modifies on memento structure TODO: complete the manage server functionality

+++ geoplatform-services/geoplatform-ws/geoplatform-ws-api/src/main/java/org/geosdi/geoplatform/responce/ServerDTO.java
@@ -56,7 +56,7 @@ import org.geosdi.geoplatform.core.model.GeoPlatformServer;
 @XmlType(propOrder = {"id", "serverUrl", "name", "alias", "layerList"})
 public class ServerDTO {
 
-    private long id;
+    private Long id;
     private String serverUrl;
     private String name;
     private String alias;
@@ -79,7 +79,7 @@ public class ServerDTO {
     /**
      * @return the id
      */
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
@@ -87,7 +87,7 @@ public class ServerDTO {
      * @param id
      *          the id to set
      */
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -143,7 +143,7 @@ public class ServerDTO {
      */
     @Override
     public String toString() {
-        return "ServerDTO{" + "id=" + id + ", serverUrl=" + serverUrl
+        return "ServerDTO{" + "id=" + id.longValue() + ", serverUrl=" + serverUrl
                 + ", alias=" + alias + ", name=" + name + '}';
     }
